Remove support for group lookup
This should be the job of another library, and the caller can just pass
in the looked up GID.

Since go-plugin-helpers uses this library, it also means the a plugin
must have `/etc/group` in it's rootfs with an entry in it.
Not all that bad, but a major pain when building a static binary to be
used as a plugin.

In any case, the lookup is only correct by luck, it's better to have the
caller of this library look up the gid as makes sense for the platform
they are on.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>