Add TLS 1.3 to list of TLS versions for Go 1.13 and up
This const is available in Go 1.12 and up (although TLS1.3 is opt-in
for that version), so this patch gates the new version based on Go version.

From the Go 1.13 release notes https://golang.org/doc/go1.13#tls_1_3

> As announced in Go 1.12, Go 1.13 enables support for TLS 1.3 in the crypto/tls
> package by default. It can be disabled by adding the value tls13=0 to the GODEBUG
> environment variable. The opt-out will be removed in Go 1.14.

Given that the implementation on Go 1.12 was not yet "stable" and required use
of the `GODEBUG` variable, we only add it for Go 1.13 and up here.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>