socket: try removing if socket-location is a directory on Linux
Due to race-conditions between containers starting and the Docker
remote API being up, containers bind-mounting the docker-socket may
cause the socket-path to be created as a directory.

This patch will attempt to remove the directory in such situations.
Removing will fail if the directory is not empty.

MacOS does not allow us to detect that the path is a directory, and we'll
return immediately instead of retrying.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>