socket: restore deprecated Transport.Dial (for now)
This restores the deprecated Transport.Dial, which were removed in commits:

- 61039d0834e53e8ae181e064229a9e79d7cd77aa (Replace deprecated Transport.Dial with Transport.DialContext)
- fb772cf57ab026bb75d4b3808e9b1761cd66e6c4 (Fix problems introduced by 61039d0)

While we should still look at removing these, the moby code currently looks to
be depending on their behavior. Removing them caused CI to fail, which blocked
us from updating to the current version of this package.

With those changes, Windows clients in CI were connecting with the wrong daemon,
causing CI failures:

    Failed
    failed to get info from daemon: Error response from daemon: client version 1.41 is too new. Maximum supported API version is 1.40
    exit status 1
    FAIL	github.com/docker/docker/integration/build	0.219s

More details on https://github.com/moby/moby/pull/41042 and https://github.com/moby/moby/pull/41084

This patch restores the deprecated parts, but keeps the new variants as well,
so that we can perform the migration in Moby when possible (after which they
can be removed again)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>