tests: Skip tests failing on darwin
TestConfigServerExclusiveRootPools and TestConfigClientExclusiveRootPools
are failing on darwin with the same error message as on Windows:

> Unable to verify certificate 1: x509: certificate signed by unknown authority

The `(*Certificate).Verify()` method from `crypto/x509` special-case
windows, darwin and ios GOOS to use a OS-specific verification process.
This process seems to consider root CAs as invalid for some unknown
reasons.

This should be further investigated.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>