tlsconfig: align client and server defaults, remove weak CBC ciphers
These ciphers were split between server and client in [docker@9b43f5a]
(Docker v1.8.0, Jun 11, 2015);

> removing the CBC ciphers from the client preferred TLS cipher suites.
> This will allow a future version of the server to also remove the CBC
> ciphers from the accepted list.
>
> This changes the server default to client + additional CBC cipher list,
> and client default to the non-CBC ciphers.

That change allowed phasing out the use of these ciphers in the client,
but (for backward-compatibility with older clients) the daemon to still
accept old ones.

Given that no current client versions should still be using these, we
should be able to remove them from the list of ciphers that are supported
by the daemon.

Now that client and server are the same, we can also use a single implementation
for both.

[docker@9b43f5a]: https://github.com/docker/docker/commit/9b43f5a3a1c12722a7f7b5ea60c28d02bac60d1c

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>