sockets: ConfigureTransport: prevent idle connections leaking FDs
Set default `MaxIdleConns` / `IdleConnTimeout` to prevent idle connections
leaking FDs.

When using the client (initialized via `dockercli` in my case) in a
long-running process, the idle connections are not released. If you
create multiple clients the FD count keeps growing indefinitely.

This can be demonstrated against Docker Desktop; other cases have not
been tested, but it also depends on the server's behavior and when the
server will drop the connections.

Other possible fixes could be:

- initialize transport from DefaultTransport (that has better defaults)
- Somehow wrap transport so all clients use the same keepalive pool
- Make this configurable. I would also need dockerCLI update then.
- Explore server side
- There is a call to `CloseIdleConns` but that lifecycle is tricky to
  manage when initializing clients via `dockerCLI` based on multiple
  configs. Possibly defining a `Finalizer` could also result in the
  desired behavior.

This patch migrates the code from [moby@5c72a95] to this module.

[moby@5c72a95]: https://github.com/moby/moby/commit/5c72a95a30bb1a1c5cdb2ad555dd1bc0622cf870

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>