nat: ParsePortSpec: validate port before proto
commit f38822229cbe4b84304ffb166160b6a253938009 made an optimization
to handle validating the proto before parsing port-numbers, but
as a result changed the error returned if the port was missing.

This patch moves the check for empty container ports to the start
to keep the old behavior. The error-message(s) can probably be
improved a bit, but this keeps the existing output.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>