Set correct SECURITY_INFORMATION bit
Prior to this change we were setting `OWNER_SECURITY_INFORMATION`
along with `DACL_SECURITY_INFORMATION`.

Since we are not touching the owner, I don't think we need to
set this.

Furthermore, the test was failing because the DACL was actually
protected (`D:P(A;;GA;;;BA)(A;;GA;;;SY)`) and we need to tell
the API about it.

Setting `PROTECTED_DACL_SECURITY_INFORMATION` fixes the issue
and allows socket creation with the required security descriptor.

Signed-off-by: Craig Gumbley <craiggumbley@gmail.com>